refactor: change project structure according to PR recomendations
