+++ src/components/CardsList.js
@@ -1,7 +1,8 @@
-import { cardDeck, shuffleDeck, getFirst21Items } from "../utils/cardHandler";
+import { cardDeck, shuffleDeck } from "../entity/cardDeck";
+import { getArrayItems } from "../utils/arrayItems";
 const CardsList = () => {
   const shuffledDeck = shuffleDeck(cardDeck);
-  const first21ShuffledCards = getFirst21Items(shuffledDeck);
+  const first21ShuffledCards = getArrayItems(shuffledDeck);
   return (
     <ul>
       {first21ShuffledCards.map((card) => (
