+++ src/utils/arrayItems.js
@@ -1,0 +1,5 @@
+function getArrayItems(items) {
+  return items.slice(0, 21);
+}
+
+export { getArrayItems };
